Feature: Getting User Information and File information
Feature: Getting User Information
- Getting User List
- UserList includes username, uid, gid (Joined Group lsit), Last login date, Last Login IP, Last password changed date
- Using /var/log/wtmp

Feature: Getting file informations
- Check file information (Permissions, Ownership, Group)
- Using stat
- FileList that need to be checked is defined in "0_fileList.h"
- Check Permissions, Ownership and Group, and edit these properties suitablely and automatically, at once.

[Kinds of UserInfo struct]
- UserName
- UID
- The number of Group
- Gruop GID List
- Last login Date
- Last changed password Date
- Last Login IP

[Kinds of FileInfo struct]
- Full path of a file
- file name
- checked (Items that need to be changed)
- OwnerUID (Existing UID, UID to be changed)
- Group GID (Existing UID, UID to be changed)
- Permissions (Existing UID, UID to be changed)

+++ hw_info.h
@@ -3,30 +3,30 @@
 #define HW_INFO_H
 #include "zz_struct.h"
 
-/* Functions that get server information (System Parts include CPU, Memory, PSU, Network, and so on.) */
+/* List of functions that get server information (System Parts include CPU, Memory, PSU, Network, and so on.) */
 void get_System_Information_from_Omreport(SystemInfo* systemBuf);
 void get_Server_Information_from_Omreport(char* hostname, char* serverModel, char* serviceTag, char* serviceCode);
 void get_CPU_Information_from_Omreport(CPUInfo* cpuBuf);
 void get_Memory_Information_from_Omreport(MEMInfo* memoryBuf);
 void get_CMOS_Battery_Status_from_Omreport(CMOS_BAT_INFO* cmosBatteryBuf);
 void get_Fan_Status_from_Omreport(FANInfo* fanBuf);
-void get_Physical_IFA_Information_from_Omreport(PHYSICAL_IFA_Info** ifaBuf, short* ifaCount);
-void get_Interface_Speed_from_Omreport(PHYSICAL_IFA_Info* ifaBuf, short ifaCount);
-void get_PSU_Status_from_Omreport(short* psuStatusBuf);
+void get_Physical_IFA_Information_from_Omreport(PHYSICAL_IFA_Info** ifaBuf, int* ifaCount);
+void get_Interface_Speed_from_Omreport(PHYSICAL_IFA_Info* ifaBuf, int ifaCount);
+void get_PSU_Status_from_Omreport(int* psuStatusBuf);
 
-/* Functions that get disk information */
-short get_VDisk_Information_from_Perccli(VDInfo** vdBuf, short* virtualDriveCnt);
-void get_VDisk_FileSystem_from_Perccli(VDInfo* vdBuf, short virtualDiskCnt);
-short get_Disk_Information_from_Perccli(DiskInfo** diskBuf, short* diskCount);
-void get_Disk_Product_Name_from_Perccli(DiskInfo* diskBuf, short diskCount);
+/* List of functions that get disk information */
+int get_VDisk_Information_from_Perccli(VDInfo** vdBuf, int* virtualDriveCnt);
+void get_VDisk_FileSystem_from_Perccli(VDInfo* vdBuf, int virtualDiskCnt);
+int get_Disk_Information_from_Perccli(DiskInfo** diskBuf, int* diskCount);
+void get_Disk_Product_Name_from_Perccli(DiskInfo* diskBuf, int diskCount);
 
-/* Functions that get HBA Card information */
-short get_HBA_Information_from_Perccli(HBAInfo* HBABuf);
+/* List of functions that get HBA Card information */
+int get_HBA_Information_from_Perccli(HBAInfo* HBABuf);
 void get_BBU_Information_from_Perccli(BBUInfo* BBUBuf);
 
-/* Functions that get CPU information from Jiffies (/proc/cpuinfo) */
+/* List of functions that get CPU information from Jiffies (/proc/cpuinfo) */
 void get_CPU_Usage_Percent_of_each_Core(float** usage_buf);
-short get_Physical_CPU_Count(short totalCore);
-short get_CPU_Core_Count();
+int get_Physical_CPU_Count(int totalCore);
+int get_CPU_Core_Count();
 
 #endif
